Modify the default options for this role
By default we don't want anything to be installed and we
don't need a brewfile to be used. Additionally a new variable
for the logged in user is created. This is so homebrew can be
properly installed using both the root user and the regular user
logged into the machine.